fix(movie null):Se arreglo la pelicula que llegaba nula en details
