+++ lib/config/router/app_router.dart
@@ -1,23 +1,52 @@
 import 'package:go_router/go_router.dart';
+import 'package:cinemapedia/presentation/views/views.dart';
 import 'package:cinemapedia/presentation/screens/screens.dart';
 
 final appRouter = GoRouter(
   initialLocation: '/',
   routes: [
-    GoRoute(
-        path: '/',
-        name: HomeScreen.name,
-        builder: (context, state) => const HomeScreen(),
-        routes: [
-          GoRoute(
-            path: 'movie/:id',
-            name: MovieScreen.name,
-            builder: (context, state) {
-              final movieId = state.pathParameters['id'] ?? 'no-id';
+    StatefulShellRoute.indexedStack(
+      builder: (context, state, navigationShell) =>
+          HomeScreen(childView: navigationShell),
+      branches: <StatefulShellBranch>[
+        StatefulShellBranch(
+          routes: [
+            GoRoute(
+              path: '/',
+              builder: (context, state) {
+                return const HomeView();
+              },
+              routes: [
+                GoRoute(
+                  path: 'movie/:id',
+                  name: MovieScreen.name,
+                  builder: (context, state) {
+                    final movieId = state.pathParameters['id'] ?? 'no-id';
 
-              return MovieScreen(movieId: movieId);
-            },
-          ),
-        ]),
+                    return MovieScreen(movieId: movieId);
+                  },
+                ),
+              ],
+            ),
+          ],
+        ),
+        StatefulShellBranch(
+          routes: [
+            GoRoute(
+              path: '/categories',
+              builder: (context, state) => const CategoriesView(),
+            )
+          ],
+        ),
+        StatefulShellBranch(
+          routes: [
+            GoRoute(
+              path: '/favorites',
+              builder: (context, state) => const FavoritesView(),
+            )
+          ],
+        )
+      ],
+    )
   ],
 );
