+++ lib/presentation/screens/movies/home_screen.dart
@@ -1,108 +1,23 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import 'package:cinemapedia/presentation/widgets/widgets.dart';
-import 'package:cinemapedia/presentation/providers/providers.dart';
+import 'package:go_router/go_router.dart';
 
 class HomeScreen extends StatelessWidget {
   static const name = 'home-screen';
 
-  const HomeScreen({super.key});
+  final StatefulNavigationShell childView;
+
+  const HomeScreen({
+    super.key,
+    required this.childView,
+  });
 
   @override
   Widget build(BuildContext context) {
-    return const Scaffold(
-      body: _HomeView(),
-      bottomNavigationBar: CustomBottomNavigation(),
-    );
-  }
-}
-
-class _HomeView extends ConsumerStatefulWidget {
-  const _HomeView();
-
-  @override
-  _HomeViewState createState() => _HomeViewState();
-}
-
-class _HomeViewState extends ConsumerState<_HomeView> {
-  @override
-  void initState() {
-    super.initState();
-    ref.read(nowPlayingMoviesProvider.notifier).loadNextPage();
-    ref.read(popularMoviesProvider.notifier).loadNextPage();
-    ref.read(upComingMoviesProvider.notifier).loadNextPage();
-    ref.read(voteAverageMoviesProvider.notifier).loadNextPage();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final initialLoading = ref.watch(initialLoadingProvider);
-
-    if (initialLoading) return const FullScreenLoader();
-
-    final slideShowMovies = ref.watch(moviesSlideshowProvider);
-    final nowPlayingMovies = ref.watch(nowPlayingMoviesProvider);
-    final popularMovies = ref.watch(popularMoviesProvider);
-    final upComingMovies = ref.watch(upComingMoviesProvider);
-    final avegareMovies = ref.watch(voteAverageMoviesProvider);
-
-    return CustomScrollView(
-      slivers: [
-        const SliverAppBar(
-          floating: true,
-          flexibleSpace: FlexibleSpaceBar(
-            title: CustomAppbar(),
-          ),
-        ),
-        SliverList(
-          delegate: SliverChildBuilderDelegate(
-            (context, index) {
-              return Column(
-                children: [
-                  //const CustomAppBar(),
-                  MoviesSlideshow(
-                    movies: slideShowMovies,
-                  ),
-                  MovieHorizontalListView(
-                    movies: nowPlayingMovies,
-                    title: 'En cines',
-                    subTitle: 'Lunes 20',
-                    loadNextPage: () => ref
-                        .read(nowPlayingMoviesProvider.notifier)
-                        .loadNextPage(),
-                  ),
-                  MovieHorizontalListView(
-                    movies: upComingMovies,
-                    title: 'Proximamente',
-                    subTitle: 'En este mes',
-                    loadNextPage: () => ref
-                        .read(upComingMoviesProvider.notifier)
-                        .loadNextPage(),
-                  ),
-                  MovieHorizontalListView(
-                    movies: popularMovies,
-                    title: 'Populares',
-                    subTitle: 'En este mes',
-                    loadNextPage: () =>
-                        ref.read(popularMoviesProvider.notifier).loadNextPage(),
-                  ),
-                  MovieHorizontalListView(
-                    movies: avegareMovies,
-                    title: 'Mejor calificadas',
-                    subTitle: 'Desde siempre',
-                    loadNextPage: () => ref
-                        .read(voteAverageMoviesProvider.notifier)
-                        .loadNextPage(),
-                  ),
-                  const SizedBox(height: 10)
-                ],
-              );
-            },
-            childCount: 1,
-          ),
-        ),
-      ],
+    return Scaffold(
+      body: childView,
+      bottomNavigationBar: CustomBottomNavigation(currentChild: childView),
     );
   }
 }
