+++ lib/presentation/views/home_views/categories_view.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class CategoriesView extends StatelessWidget {
+  const CategoriesView({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('Categories View'),
+      ),
+      body: const Center(
+        child: Text('Categorías'),
+      ),
+    );
+  }
+}
