+++ lib/presentation/widgets/shared/custom_bottom_navigation.dart
@@ -1,23 +1,35 @@
 import 'package:flutter/material.dart';
+import 'package:go_router/go_router.dart';
 
 class CustomBottomNavigation extends StatelessWidget {
-  const CustomBottomNavigation({super.key});
+  final StatefulNavigationShell currentChild;
+  const CustomBottomNavigation({
+    super.key,
+    required this.currentChild,
+  });
 
   @override
   Widget build(BuildContext context) {
     return BottomNavigationBar(
-      elevation: 0,
+      currentIndex: currentChild.currentIndex,
+      onTap: (value) => currentChild.goBranch(value),
       items: const [
         BottomNavigationBarItem(
-          icon: Icon(Icons.home_max),
+          icon: Icon(
+            Icons.home_max,
+          ),
           label: 'Inicio',
         ),
         BottomNavigationBarItem(
-          icon: Icon(Icons.label_outline),
-          label: 'Categorias',
+          icon: Icon(
+            Icons.label_outline,
+          ),
+          label: 'Categorías',
         ),
         BottomNavigationBarItem(
-          icon: Icon(Icons.favorite_outline),
+          icon: Icon(
+            Icons.favorite_outline,
+          ),
           label: 'Favoritos',
         )
       ],
